+++ swiyu-core-business-service-app/src/main/java/ch/admin/bj/swiyu/core/business/modules/dataimport/service/DemoDataAsyncExecutor.java
@@ -28,6 +28,7 @@ public class DemoDataAsyncExecutor {
                 setSystemSecurityContext();
 
                 demoDataImportService.generateBusinessPartners();
+                demoDataImportService.generateBusinessPartnerIdentities();
                 demoDataImportService.generateIdentifierEntries();
                 demoDataImportService.deleteDemoTrustOnboardingSubmissions();
                 demoDataImportService.generateTrustOnboardingSubmissions();